refactor: 전역 예외 클래스 web으로 이동
